pyproject.toml
[project]
name = "optivox-backend"
version = "4.0.0"
description = "OptiVox DB — Agentic AI SQL Studio powered by CrewAI + LiteLLM + LanceDB"
requires-python = ">=3.11"

dependencies = [
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.1",
    # Pydantic
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # CrewAI + LiteLLM (Gemini via LiteLLM)
    "crewai[google-genai]>=1.10.0",
    "litellm>=1.40.0",
    # Database
    "sqlalchemy>=2.0.30",
    "pymysql>=1.1.1",
    "psycopg2-binary>=2.9.9",
    "oracledb>=2.3.0",
    # SQL parsing / validation
    "sqlglot>=25.0.0",
    # RAG — Phase 2
    "lancedb>=0.11.0",
    "sentence-transformers>=3.0.0",
    "pandas>=2.2.0",
    # Utilities
    "cachetools>=5.3.3",
    "pyarrow>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0",
    "httpx>=0.27.0",
]
